# views/dist/swagger.yaml
openapi: 3.0.0
...
servers:
  - description: SwaggerHub API Auto Mocking
    url: http://localhost:3000
info:
  description: Test APIs document Swagger
  version: "1.0.0"
  title: Thu vien sach APIs document Swagger
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
basePath: "/v2"
...
paths:
  /api/product/database:
    get:
      tags:
        - database
      summary: "Show all product database"
      description: "Returns all json products"
      operationId: "productDatabase"
      produces:
        - "application/xml"
        - "application/json"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/components/schemas/Clothes"
        "400":
          description: "Errors"
        "404":
          description: "Error"
...
components:
  schemas:
    Clothes:
      type: object
      properties:
        name:
          type: string
        desc:
          type: string
        price:
          type: string
        imageUrl:
          type: string
        color:
          type: string
        colorPicker:
          type: array
          items:
            type: string
        category:
          type: string
        type:
          type: string
        sale:
          type: boolean
          default: false
        imageItem:
          type: array
          items:
            type: string
        date:
          type: string
      xml:
        name: User
...
  securitySchemes:
    test_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: "http://test-swagger.herokuapp.com"
          scopes:
            "write:users": modify your account
            "read:users": read your information
    api_key:
      type: apiKey
      name: api_key
      in: header